Implement DeleteBook icon in book details view
In single pane layout, delete book when icon is clicked in details view. Also, improve the way the
app determines if we have a single or dual pane layout. Now using a bool value in the .xml and this
value is read from a static IsTwoPaneLayout() utility method I created.