Change the way the book details view is updated
Dual-pane: now we only update the book details view when a book was displayed in the AddBook view.
This happens when a book is read from the DB or downloaded from server. Otherwise, no need to
update the details view since we want to display the same book that was being displayed before
the AddBook view was opened.